Fix failed CI job
It seems, in earlier versions of *Crystal*, `private class`es are
not really isolated. This does not occur with v1.5.1 of *Crystal*,
as the CI passes for that version.

Renaming the classes makes the CI pass, but I took a different
approach of permitting all columns on the operations, giving the
classes a similar *signature*.